Using broadcasting instead of repeat_interleave
This speeds up the forward method (therefore the inference) of our model, however training speed remains the same,
because the backwards call now takes the exact time originally gained longer.
But since we are more interesse in the inference this is no problemo.
Also added some docstrings and comments to the ColBERT class, however many docstrings and typ-hints are still missing and should be added in the near future.